Added for displayRecord method in Crudify trait
